Refactor `relvar-core/src/query/mod.rs` into multiple files using the Facade pattern.
Splits the monolithic query module into `ast.rs`, `builder.rs`, `execute.rs`, and `tests.rs` to enforce better separation of concerns and reduce file bloat.
